+++ src/app/pages/auth/sign-up/signup.component.html
@@ -9,112 +9,7 @@
          
         </div>
         <div class="login-main">
-          <form (ngSubmit)="handleSignup($event)" autocomplete="off" #frm="ngForm">
-            <div class="mb-3 text-center">
-              <h4>Registro</h4>
-            </div>
-
-            <!-- Google sign-up button -->
-            <div class="form-group mb-3">
-              <button type="button" class="google-btn w-100">
-                <span class="google-icon">G</span>
-                Registrar con Google
-              </button>
-            </div>
-
-            <!-- Separator -->
-            <div class="d-flex align-items-center mb-3">
-              <hr class="flex-grow-1" />
-              <span class="mx-2">o</span>
-              <hr class="flex-grow-1" />
-            </div>
-
-            <div class="d-flex align-items-center mb-3">
-              <div class="form-group px-1">
-                <label for="name" class="form-label mb-2">
-                  Nombre
-                </label>
-                <input #name="ngModel" type="text" class="form-control" id="name" name="name"
-                  placeholder="Ingrese su nombre" [(ngModel)]="user.name" [ngClass]="{
-                    'is-invalid': name.invalid && name.touched,
-                    'is-valid': name.valid && name.touched
-                  }" required />
-                <div *ngIf="name.invalid" class="invalid-feedback">
-                  Por favor ingrese un nombre valido
-                </div>
-              </div>
-              <div class="form-group px-1">
-                <label for="lastname" class="form-label mb-2">
-                  Apellido
-                </label>
-                <input #lastname="ngModel" type="text" class="form-control" id="lastname" name="lastname"
-                  placeholder="Ingrese su apellido" [(ngModel)]="user.lastname" [ngClass]="{
-                    'is-invalid': lastname.invalid && lastname.touched,
-                    'is-valid': lastname.valid && lastname.touched
-                  }" required />
-                <div *ngIf="lastname.invalid" class="invalid-feedback">
-                  Por favor ingrese un apellido valido
-                </div>
-              </div>
-            </div>
-
-            <div class="d-flex align-items-center mb-3">
-              <div class="form-group px-1 flex-grow-1">
-                <label for="email" class="form-label mb-2">
-                  Email
-                </label>
-                <input #email="ngModel" type="email" class="form-control" id="email" name="email"
-                  placeholder="ejemplo.email@gmail.com" [(ngModel)]="user.email" [ngClass]="{
-                    'is-invalid': email.invalid && email.touched,
-                    'is-valid': email.valid && email.touched
-                  }" required />
-                <div *ngIf="email.invalid" class="invalid-feedback">
-                  Por favor ingrese un email valido
-                </div>
-              </div>
-              <div class="form-group px-1 flex-grow-1">
-                <label for="age" class="form-label mb-2">
-                  Edad
-                </label>
-                <input #age="ngModel" type="number" class="form-control" id="age" name="age"
-                  placeholder="-" [(ngModel)]="user.age" [ngClass]="{
-                    'is-invalid': age.invalid && age.touched,
-                    'is-valid': age.valid && age.touched
-                  }" required />
-                <div *ngIf="age.invalid" class="invalid-feedback">
-                  Por favor ingrese una edad valida
-                </div>
-              </div>
-            </div>
-
-            <div class="form-group mb-5">
-              <label for="password" class="form-label mb-2">Contraseña</label>
-              <div class="form-input position-relative">
-                <input #password="ngModel" type="password" class="form-control" id="password" name="password"
-                placeholder="Mínimo 8 caracteres" [(ngModel)]="user.password" [ngClass]="{
-                  'is-invalid': password.invalid && password.touched,
-                  'is-valid': password.valid && password.touched
-                }" required />
-                <div *ngIf="password.invalid" class="invalid-feedback">
-                 Por favor ingrese una contraseña valida
-                </div>
-              </div>
-            </div>
-
-            <div class="form-group mb-0">
-              <div class="text-center">
-                <button class="btn btn-primary" type="submit">
-                  Registrar
-                </button>
-              </div>
-              @if(signUpError) {
-                <p class="text-danger">{{signUpError}}</p>
-              } @else if (validSignup){
-                <p class="text-success">Usuario registrado, por favor ingrese a login</p>
-              }
-            </div>
-            <p class="mt-3 text-center">Ya tienes una cuenta? <a routerLink="/login">Login</a></p>
-          </form>
+          <app-sign-up-form></app-sign-up-form>
         </div>
       </div>
     </div>
