+++ src/app/pages/auth/sign-up/signup.component.ts
@@ -1,55 +1,15 @@
-import { CommonModule } from '@angular/common';
-import { Component, ViewChild } from '@angular/core';
-import { FormsModule, NgModel } from '@angular/forms';
-import { Router, RouterLink } from '@angular/router';
-import { AuthService } from '../../../services/auth.service';
-import { IUser } from '../../../interfaces';
+import { Component } from '@angular/core';
+import { RouterLink } from '@angular/router';
 import { LogoTituloComponent } from '../../../components/logo-titulo/logo-titulo.component';
+import { SignUpFormComponent } from '../../../components/auth/sign-up/sign-up-form/sign-up-form.component';
 
 @Component({
   selector: 'app-signup',
   standalone: true,
-  imports: [CommonModule, FormsModule, RouterLink,LogoTituloComponent],
+  imports: [RouterLink, LogoTituloComponent, SignUpFormComponent],
   templateUrl: './signup.component.html',
   styleUrl: './signup.component.scss'
 })
 export class SigUpComponent {
-  public signUpError!: String;
-  public validSignup!: boolean;
-  @ViewChild('name') nameModel!: NgModel;
-  @ViewChild('lastname') lastnameModel!: NgModel;
-  @ViewChild('email') emailModel!: NgModel;
-  @ViewChild('age') ageModel!: NgModel;
-  @ViewChild('password') passwordModel!: NgModel;
-
-  public user: IUser = {};
-
-  constructor(private router: Router, 
-    private authService: AuthService
-  ) {}
-
-  public handleSignup(event: Event) {
-    event.preventDefault();
-    if (!this.nameModel.valid) {
-      this.nameModel.control.markAsTouched();
-    }
-    if (!this.lastnameModel.valid) {
-      this.lastnameModel.control.markAsTouched();
-    }
-    if (!this.emailModel.valid) {
-      this.emailModel.control.markAsTouched();
-    }
-    if (!this.ageModel.valid) {
-      this.ageModel.control.markAsTouched();
-    }
-    if (!this.passwordModel.valid) {
-      this.passwordModel.control.markAsTouched();
-    }
-    if (this.emailModel.valid && this.passwordModel.valid) {
-      this.authService.signup(this.user).subscribe({
-        next: () => this.validSignup = true,
-        error: (err: any) => (this.signUpError = err.description),
-      });
-    }
-  }
+  constructor() {}
 }
